Introduce PredicateExpectation + vital refactoring
Rename static ExpectationTranslators into ExpectationFactory interface and its implementation - DefaultExpectationFactory.
Similar thing for ActualTranslators.
There is FacadeExpectationFactory instead of plain FacadeExpectation constructor. This is to be able to inject ExpectationFactory and ActualFactory.
There is another implementation of ExpectationFactory - SpockExpectationFactory which adds additional IF-ELSE case to map Groovy's Closure type to a PredicateExpectation.
For now, PredicateExpectation is tested only in FacadeExpectationForGroovySpec - in which you setup PredicateExpectation using Groovy's Closure.